Large set of changes to more fully implement mit-scheme's spec and add more convenient functionality
Appease whiners and change the comment style to two semi-colons and a space.
Make all symbols lower case for case-sensitive readers.
Allow the reporter for a condition to be a string, and to not specify a reporter.
signal-condition now continues the search for condition handlers if a handler
returns.
Add a set of condition signalers for specific condition-types.
Export condition-type/name, which was left out by accident.
Add a set of procedures to invoke restarts that adhere to a standard protocol.
Add illegal-datum, datum-out-of-range, wrong-type-datum, wrong-type-argument,
bad-range-argument, control-error, and no-such-restart condition types.
Fix report to print more friendly strings.